chore: refactor auto beta release to be publish branch
