--- .github/workflows/beta-release.yaml
@@ -1,63 +1,0 @@
-name: Beta Release
-
-on:
-    workflow_call:
-
-jobs:
-    beta_release:
-        runs-on: ubuntu-latest
-        env:
-            NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
-            NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
-        steps:
-            - uses: actions/checkout@v4
-              with:
-                  ref: ${{ github.head_ref }}
-            - uses: actions/setup-node@v4
-              with:
-                  node-version: "18.x"
-                  registry-url: "https://registry.npmjs.org"
-                  cache: "yarn"
-
-            - name: Install dev dependencies
-              run: yarn install
-
-            - name: Setup npm credentials file
-              run: echo "//registry.npmjs.org/:_authToken=$NODE_AUTH_TOKEN" >> .npmrc
-
-            - name: Setup git credentials
-              run: |
-                  git config --global user.name 'Auto Release Bot'
-                  git config --global user.email 'auto-release-bot@users.noreply.github.com'
-
-            - name: Get current package.json version
-              run: echo "PACKAGE_VERSION=$(npm pkg get version)" >> $GITHUB_ENV
-
-            - name: Setup Beta Release Version
-              run: node beta-release.js --issue $GITHUB_PR_NUMBER
-              env:
-                  GITHUB_PR_NUMBER: ${{ github.event.pull_request.number }}
-
-            - name: Release a new beta version
-              run: npm publish --tag beta
-              env:
-                  NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
-
-            - uses: actions/github-script@v6
-              with:
-                  script: |
-                      github.rest.issues.createComment({
-                        issue_number: context.issue.number,
-                        owner: context.repo.owner,
-                        repo: context.repo.repo,
-                        body: `Beta version released with the last commit 🚀
-
-                        \`\`\`
-                        yarn add pixi.js@${{ env.NEW_VERSION }}
-                        \`\`\`
-                        or
-                        \`\`\`
-                        npm install pixi.js@${{ env.NEW_VERSION }}
-                        \`\`\`
-                        `
-                      })
